types: export types from Linux

+++ include/string.h
@@ -21,14 +21,14 @@
 
 #include <types.h>
 
-static inline void memcpy(uint8_t *dest, const uint8_t *src, uint32_t len)
+static inline void memcpy(uint8_t *dest, const uint8_t *src, size_t len)
 {
 	for (; len != 0; len--) {
 		*dest++ = *src++;
 	}
 }
 
-static inline void memset(void *dest, uint8_t val, uint32_t len)
+static inline void memset(void *dest, uint8_t val, size_t len)
 {
 	uint8_t *dst = (uint8_t *)dest;
 
@@ -37,7 +37,7 @@ static inline void memset(void *dest, uint8_t val, uint32_t len)
 	}
 }
 
-static inline void bzero(void *dest, uint32_t len)
+static inline void bzero(void *dest, size_t len)
 {
 	memset(dest, 0, len);
 }
